server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
ConfigArgParse==1.7
dnspython==2.7.0
docopt==0.6.2
fastapi==0.115.6
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
flatbuffers==24.12.23
fsspec==2025.2.0
gevent==24.11.1
geventhttpclient==2.3.3
googleapis-common-protos==1.70.0
greenlet==3.1.1

# --- Core RPC stack (keep grpcio & grpcio-tools in lockstep; protobuf exact) ---
grpcio==1.75.0
grpcio-health-checking==1.75.0
protobuf==6.32.1
grpcio-reflection==1.75.0

h11==0.14.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
locust==2.33.1
MarkupSafe==3.0.2
motor==3.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.2
onnx==1.17.0
onnxruntime==1.20.1
packaging==24.2
pipreqs==0.4.13
psutil==7.0.0
pydantic==2.10.5
pydantic_core==2.27.2
pymongo==4.11.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.3.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scipy==1.15.1
sentencepiece==0.2.0
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
yarg==0.1.10
zope.event==5.0
zope.interface==7.2